JWT securisation + front integration
JWT controller check, front end api call with authorization header included and Redux + localStorage for token storing

+++ back-office/server/src/main/java/proj/cloud/ath/controllers/api/v1/BrandController.java
@@ -8,59 +8,101 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import proj.cloud.ath.entities.Brand;
 import proj.cloud.ath.response.RestApiResponse;
 import proj.cloud.ath.services.BrandService;
+import proj.cloud.ath.utils.JwtUtil;
 
 @RestController
 @RequestMapping("/api/v1/brand")
 @CrossOrigin(origins = { "http://localhost:3000" })
 public class BrandController {
+
     @Autowired
     private BrandService service;
 
+    @Autowired
+    private JwtUtil jwtUtil;
+
     @GetMapping
-    public RestApiResponse findAll() {
-        return new RestApiResponse(service.findAll(), 200);
+    public RestApiResponse findAll(@RequestHeader(value = "Authorization") String bearerToken) {
+        String token = bearerToken.substring(7);
+        RestApiResponse response = new RestApiResponse();
+        if (jwtUtil.isValidToken(token)) {
+            response = new RestApiResponse(service.findAll(), 200);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
+        return response;
     }
 
     @GetMapping("/{id}")
-    public RestApiResponse findById(@PathVariable Long id) {
-        return new RestApiResponse(service.findById(id), 200);
+    public RestApiResponse findById(@RequestHeader(value = "Authorization") String bearerToken, @PathVariable Long id) {
+        String token = bearerToken.substring(7);
+        RestApiResponse response = new RestApiResponse();
+        if (jwtUtil.isValidToken(token)) {
+            response = new RestApiResponse(service.findById(id), 200);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
+        return response;
     }
 
     @PostMapping
-    public RestApiResponse create(@RequestBody Brand brand) {
+    public RestApiResponse create(
+            @RequestHeader(value = "Authorization") String bearerToken, @RequestBody Brand brand) {
+        String token = bearerToken.substring(7);
         RestApiResponse response = new RestApiResponse();
-        brand.setState(0);
-        service.save(brand);
-        response.setPayload(brand);
-        response.setStatus(201);
+        if (jwtUtil.isValidToken(token)) {
+            brand.setState(0);
+            service.save(brand);
+            response.setPayload(brand);
+            response.setStatus(201);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
         return response;
     }
 
     @PutMapping
-    public RestApiResponse update(@RequestBody Brand brand) {
+    public RestApiResponse update(@RequestHeader(value = "Authorization") String bearerToken,
+            @RequestBody Brand brand) {
+        String token = bearerToken.substring(7);
         RestApiResponse response = new RestApiResponse();
-        response.setStatus(400);
-        Brand oldBrand = service.findById(brand.getId());
-        if (oldBrand != null) {
-            oldBrand.setName(brand.getName());
-            service.save(oldBrand);
-            response.setStatus(200);
-            response.setPayload(oldBrand);
+        if (jwtUtil.isValidToken(token)) {
+            response.setStatus(400);
+            Brand oldBrand = service.findById(brand.getId());
+            if (oldBrand != null) {
+                oldBrand.setName(brand.getName());
+                service.save(oldBrand);
+                response.setStatus(200);
+                response.setPayload(oldBrand);
+            }
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
         }
         return response;
     }
 
     @DeleteMapping("/{id}")
-    public RestApiResponse delete(@PathVariable Long id) {
+    public RestApiResponse delete(@RequestHeader(value = "Authorization") String bearerToken, @PathVariable Long id) {
+        String token = bearerToken.substring(7);
         RestApiResponse response = new RestApiResponse();
-        service.delete(id);
-        response.setStatus(200);
+        if (jwtUtil.isValidToken(token)) {
+            service.delete(id);
+            response.setStatus(200);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
         return response;
     }
 }
